refactor: :recycle: rename dirs
