+++ githubCrawler/crates_io/top1000.ts
@@ -1,6 +1,8 @@
+import { Database } from "bun:sqlite";
 import { Artifact } from "../artifact.type";
 
 const fetchTop1000 = async (): Promise<Artifact.Root> => {
+  const db = new Database("tmp/cratesio.sqlite", { readonly: true });
   return {};
 };
