+++ githubCrawler/npmjs_com/getBulkPackagesInfo.ts
